+++ Assets/NatShare/Plugins/Managed/INatShare.cs
@@ -11,8 +11,8 @@ namespace NatShareU {
 		bool ShareText (string text, ShareCallback callback);
 		bool ShareImage (byte[] pngData, string message, ShareCallback callback);
 		bool ShareMedia (string path, string message, ShareCallback callback);
-		bool SaveToCameraRoll (byte[] pngData);
-        bool SaveToCameraRoll (string videoPath);
+		bool SaveToCameraRoll (byte[] pngData, string album);
+        bool SaveToCameraRoll (string path, string album);
         Texture2D GetThumbnail (string videoPath, float time);
 	}
 }