+++ Assets/NatShare/Plugins/Managed/NatShareBridge.cs
@@ -29,9 +29,9 @@ namespace NatShareU.Platforms {
         [DllImport(Assembly, EntryPoint = "NSShareMedia")]
         public static extern bool Share (string path, string message);
         [DllImport(Assembly, EntryPoint = "NSSaveImageToCameraRoll")]
-        public static extern bool SaveToCameraRoll (byte[] pngData, int dataSize);
+        public static extern bool SaveToCameraRoll (byte[] pngData, int dataSize, string album);
         [DllImport(Assembly, EntryPoint = "NSSaveMediaToCameraRoll")]
-        public static extern bool SaveToCameraRoll (string path);
+        public static extern bool SaveToCameraRoll (string path, string album);
         [DllImport(Assembly, EntryPoint = "NSGetThumbnail")]
         public static extern bool GetThumbnail (string path, float time, ref IntPtr pixelBuffer, ref int width, ref int height);
         [DllImport(Assembly, EntryPoint = "NSFreeThumbnail")]
@@ -42,8 +42,8 @@ namespace NatShareU.Platforms {
         public static bool Share (string text) { return false; }
         public static bool Share (byte[] pngData, int dataSize, string message) { return false; }
         public static bool Share (string path, string message) { return false; }
-        public static bool SaveToCameraRoll (byte[] pngData, int dataSize) { return false; }
-        public static bool SaveToCameraRoll (string path) { return false; }
+        public static bool SaveToCameraRoll (byte[] pngData, int dataSize, string album) { return false; }
+        public static bool SaveToCameraRoll (string path, string album) { return false; }
         public static bool GetThumbnail (string path, float time, ref IntPtr pixelBuffer, ref int width, ref int height) { return false; }
         public static void FreeThumbnail (IntPtr pixelBuffer) {}
         #endif